requirements.txt
# requirements.txt
# ── Web framework ─────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18      # file uploads

# ── Database ──────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0               # async PostgreSQL driver
alembic==1.14.0

# ── Firebase ──────────────────────────────────────────
firebase-admin==6.5.0         # Realtime DB + Storage

# ── Thermal printer & Thai receipts ───────────────────
python-escpos==3.1.0
Pillow==11.0.0

# ── Telegram ──────────────────────────────────────────
httpx==0.27.2                 # async HTTP for Bot API calls

# ── Auth / Security ───────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT for QR tokens
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# ── Task scheduling ───────────────────────────────────
apscheduler==3.10.4           # payroll cycle, shift ping jobs

# ── Utilities ─────────────────────────────────────────
pydantic-settings==2.6.1
python-dotenv==1.0.1
structlog==24.4.0

# ── Dev / Testing ─────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2                 # test client